% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADASYN.R
\name{ADASYN}
\alias{ADASYN}
\title{ADASYN: Adaptive Synthetic Sampling}
\usage{
ADASYN(x, y, beta = 0.65, k = 5)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{beta}{Balance level (0, 1], when beta=1, the dataset is fully balanced.}

\item{k}{Numer of nearest neighbors}
}
\value{
\item{newData}{A data frame generated by ADASYN algorithm.}
}
\description{
This function implements ADASYN sampling.
}
\examples{
data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
newData<- ADASYN(x, y, beta=0.8, k=5)
}
\references{
He, H., Y. Bai, et al. (2008). "ADASYN: Adaptive synthetic sampling approach for imbalanced learning." IEEE Neural Networks(IJCNN)
}
