% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceCascade.R
\name{BalanceCascade}
\alias{BalanceCascade}
\alias{BalanceCascade.data.frame}
\title{Implementation of BalanceCascade Algorithm}
\usage{
BalanceCascade(x, ...)

\method{BalanceCascade}{data.frame}(x, y, iter = 4, ...)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{...}{Arguments to be passed to methods (see below).}

\item{y}{A vector of response variable from training data.}

\item{iter}{Number of iterations for base classifiers training.}
}
\value{
An object of class BalanceCascade, which is a list with the following components:
\item{call}{Function call}
\item{iter}{Number of iterations for base classifiers training.}
\item{classLabels}{Names of class labels.}
\item{base}{Types of base learner.}
\item{alphas}{Weights of base learners.}
\item{fits}{Fitted ensembled model.}
\item{thresh}{Threshold dor classification.}
}
\description{
This function implements BalanceCascade algorithm for binary class imbalance classification.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- BalanceCascade(x, y, iter = 4)
output<- predict(model, x)
}
\references{
X. Y. Liu, J. Wu and Z. H. Zhou. April 2009. \emph{Exploratory Undersampling for Class-Imbalance Learning.} IEEE Transactions on Systems, Man, and Cybernetics, Part B: Cybernetics, 2009, 39(2), pp. 539-550.
}
