% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceBagging.R
\name{bbagging}
\alias{bbagging}
\alias{bbagging.data.frame}
\title{BalanceBagging: Bagging based algorithm to deal with class imbalance}
\usage{
bbagging(x, ...)

\method{bbagging}{data.frame}(
  x,
  y,
  numBag = 40,
  base = treeBag,
  type = "SMOTEBagging",
  allowParallel = FALSE,
  ...
)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{numBag}{Number of bag.}

\item{base}{Base learner}

\item{type}{Type of bagging-based algorithm, including "SMOTEBagging","RUSBagging","RBBagging" and "ROSBagging".}

\item{allowParallel}{A logical number to control the parallel computing. If allowParallel = TRUE, the function is run using parallel techniques.}
}
\value{
An object of class bbag, which is a list with the following components:
\item{call}{Function call.}
\item{base}{Types of base learner.}
\item{type}{Type of bagging-based algorithm.}
\item{fits}{Fitted bagging-based model.}
}
\description{
This function implements bagging-based algorithm for imbalance classification. Four algorithms can be found in the current version: SMOTEBagging, RUSBagging, RBBagging and ROSBagging.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn, p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- bbagging(x, y, type = "SMOTEBagging", allowParallel=TRUE)
output <- predict (model, x)
}
\references{
S. Hido, H. Kashima, Y. Takahashi. Roughly balanced bagging for imbalanced data. Statistical Analysis & Data Mining, 2009, 2(5-6), pp.412-426.

S. Wang, X. Yao. 2009. Diversity analysis on imbalanced data sets by using ensemble models. IEEE Symposium on Computational Intelligence, 2009, pp. 324-331.
}
