% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceBoost.R
\name{bboost}
\alias{bboost}
\alias{bboost.data.frame}
\title{BalanceBoost - Implementation of Boost-based Algorithms}
\usage{
bboost(x, ...)

\method{bboost}{data.frame}(
  x,
  y,
  iter = 40,
  base = treeBoost,
  type = "AdaBoost",
  costRatio = 56/11,
  ...
)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{...}{Arguments to be passed to methods (see below).}

\item{y}{A vector of response variable from training data.}

\item{iter}{Number of iterations for base classifiers training.}

\item{base}{Types of base learner.}

\item{type}{Type of boosting-based algorithm, including "AdaBoost", "SMOTEBoost", "RUSBoost", "AdaC2"}

\item{costRatio}{Cost Ratio between Majority class and Minority class.}
}
\value{
An object of class bboost, which is a list with the following components:
\item{call}{Function call.}
\item{type}{Type of boosting-based algorithm.}
\item{base}{Types of base learner.}
\item{classLabels}{Names of class labels.}
\item{fits}{Fitted boosting-based model.}
\item{alpha}{Weights of base learners.}
}
\description{
This function implements boost-based algorithm for imbalance classification. Four algorithms can be found in the current version: Adaboost, SMOTEboost, RUSBoost, AdaC2.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition (Korean$Churn, p=0.75, list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- bboost(x, y, base = treeBoost, type = "AdaBoost")
output <- predict (model, x)
}
\references{
N. Chawla, A. Lazarevic, L. Hall, K.W. Bowyer. \emph{SMOTEBoost: improving prediction of the minority class in boosting}. Proceeding of PKDD, 2003, pp. 107-119.

Y. Sun, M.S. Kamel, A.K. Wong, Y. Wang. 2007. \emph{Cost-sensitive boosting for classification of imbalanced data}, Pattern Recognition, 2007, 40 (12), pp. 3358-3378.
}
