% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSC45.R
\name{CSC45}
\alias{CSC45}
\alias{CSC45.data.frame}
\title{Cost-sensitive C4.5 decision tree for binary classification}
\usage{
CSC45(x, y, pruning = TRUE, minIns = 2, costRatio = 11/56)

\method{CSC45}{data.frame}(x, y, pruning = TRUE, MDL = TRUE, minIns = 2, costRatio = 11/56)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{pruning}{A logical number to determine whether to prune the tree. If pruning=TRUE, do the pruning process.}

\item{MDL}{Prune using Minimum Description Length principle.}

\item{minIns}{Minimum number of instances for split.}

\item{costRatio}{Cost Ratio between Majority class and Minority class.}
}
\value{
(CSC45) An object with the following itens:
\item{pruning}{A logical number to indicate whether to prune the tree.}
\item{tree}{Fitted cost-sensitive C4.5 decision tree.}
\item{classLabels}{Names of class labels.}
\item{costRatio}{Ratio of misclassification cost between the majority and minority class.}
}
\description{
This function implements cost-sensitive C4.5 decision tree using an instance-weighting method.
}
\examples{
data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- CSC45(x, y, pruning = TRUE)
output <- predict (model, x)
}
\references{
M. T. Kai. An instance-weighting method to induce cost-sensitive trees. IEEE Transactions on Knowledge & Data Engineering, 14(3), 2002, pp. 659-665

Copyright (C) 2018 Bing Zhu
}
