% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyEnsemble.R
\name{EasyEnsemble}
\alias{EasyEnsemble}
\alias{EasyEnsemble.data.frame}
\title{EasyEnsemble}
\usage{
EasyEnsemble(x, ...)

\method{EasyEnsemble}{data.frame}(x, y, iter = 4, allowParallel = FALSE, ...)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{...}{Arguments to be passed to methods (see below).}

\item{y}{A vector of response variable from training data.}

\item{iter}{Number of iterations for base classifiers training.}

\item{allowParallel}{A logical number to control the parallel computing. If allowParallel = TRUE, the function is run using parallel techniques.}
}
\value{
An object of class EasyEnsemble, which is a list with the following components:
\item{call}{Function call.}
\item{iter}{Number of iterations for base classifiers training.}
\item{fits}{Fitted ensembled model.}
\item{base}{Types of base learner.}
\item{alphas}{Weights of base learners.}
\item{classlabels}{names of class labels.}
}
\description{
This function implements EasyEnsemble algorithm for binary imbalance classification.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- EasyEnsemble(x, y, allowParallel=TRUE)
output <- predict (model, x)
}
\references{
X. Y. Liu, J. Wu and Z. H. Zhou. \emph{Exploratory Undersampling for Class-Imbalance Learning.} IEEE Transactions on Systems, Man, and Cybernetics, Part B: Cybernetics, 2009, 39(2), pp. 539-550.
}
