% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InsExs}
\alias{InsExs}
\title{InsExs}
\usage{
InsExs(instance, dataknns, numExs, nomAtt)
}
\arguments{
\item{instance}{Selected instance}

\item{dataknns}{nearest instance set}

\item{numExs}{number of new instances generated for each instance}

\item{nomAtt}{indicators of factor variables}
}
\value{
\item{newIns}{Matrix of new instances.}
}
\description{
Generate Synthetic instances from nearest neighborhood
}
\keyword{internal}
