% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Korean}
\alias{Korean}
\title{Korean data documentation}
\format{
A data frame with 26224 rows and 11 variables.
}
\usage{
data(Korean)
}
\description{
Korean data documentation
}
\references{
reference
}
\author{
Name \email{email@provider}
}
\keyword{data}
