% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWMOTE.R
\name{MWMOTE}
\alias{MWMOTE}
\title{Majority Weighted Minority Oversampling Technique (MWMOTE)}
\usage{
MWMOTE(form, data, percOver = 1400, k1 = 5, k2 = 5, CThresh = 3)
}
\arguments{
\item{form}{A model formula.}

\item{data}{A data frame of training data.}

\item{percOver}{Percent of new instance generated for each minority instance. percOver/100 is the number of new instances per minority point.}

\item{k1}{Number of neighbours for filtering.}

\item{k2}{Number of neighbours for selecting majority instances.}

\item{CThresh}{Threshold to determine the number of clusters.}
}
\value{
\item{newData}{A data frame of the oversampled data using MWMOTE.}
}
\description{
This function implements MWMOTE sampling (Majority Weighted Minority Oversampling Technique).
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
newData<- MWMOTE(Churn ~., trainset)
}
\references{
Barua, S., M. Islam, et al. (2014).\emph{"MWMOTE--Majority Weighted Minority Oversampling Technique for Imbalanced Data Set Learning."} IEEE Transactions on Knowledge and Data Engineering, 26(2): 405-425.
}
