% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Numeralize}
\alias{Numeralize}
\title{Numeralize}
\usage{
Numeralize(data, form = NULL)
}
\arguments{
\item{data}{A two-dimensional dataset.}

\item{form}{Model formula.}
}
\value{
\item{numerMatrix}{Numerilized Matrix.}
}
\description{
Convert dataset into numeric matrix.
}
\keyword{internal}
