% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceCascade.R
\name{predict.BalanceCascade}
\alias{predict.BalanceCascade}
\title{Predict Method for BalanceCascade Object}
\usage{
\method{predict}{BalanceCascade}(object, x, type = "class", ...)
}
\arguments{
\item{object}{An object of BalanceCascde class.}

\item{x}{A data frame of the predictors from testing data.}

\item{type}{Types of output, which can be \strong{probability} and \strong{class} (predicted label). Default is probability.}

\item{...}{Additional arguments for predict method.}
}
\value{
Two types of output can be selected:
\item{probability}{Estimated probability of being a minority instance. The probability is averaged by using an equal-weight majority vote by all weak learners.}
\item{class}{Predicted class of the instance. Instances of probability larger than 0.5 are predicted as 1, otherwise 0.}
}
\description{
Predicting instances in test set using BalanceCascade object.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- BalanceCascade(x, y, allowParallel=TRUE)
output <- predict(model, x)
}
