% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceBagging.R
\name{predict.bbag}
\alias{predict.bbag}
\title{Predict Method for bbagging object}
\usage{
\method{predict}{bbag}(object, x, type = "class", ...)
}
\arguments{
\item{object}{An object of bbaging class.}

\item{x}{A data frame of the predictors from testing data.}

\item{type}{Types of output, which can be probability and class (predicted label). Default is probability.}

\item{...}{Additional arguments for predict method.}
}
\value{
Two types of output can be selected:
\item{probability}{Estimated probability of being a minority instance. The probability is averaged by using an equal-weight majority vote by all weak learners.}
\item{class}{Predicted class of the instance. Instances of probability larger than 0.5 are predicted as 1, otherwise 0.}
}
\description{
Predicting instances in test set using bbagging object.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x<- trainset[, -11]
y<- trainset[, 11]
model <- bbagging(x, y, type = "SMOTEBagging", allowParallel=TRUE)
output <- predict(model, x, type = "probability") # return probability estimation
output <- predict(model, x, type = "class") # return predicted class
}
