% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSC45.R
\name{predict.CSC45}
\alias{predict.CSC45}
\title{Predict Method for CSC4.5 Object}
\usage{
\method{predict}{CSC45}(object, x.test, type = "class", ...)
}
\arguments{
\item{object}{An object of CSC4.5 class.}

\item{x.test}{A data frame of the predictors from testing data.}

\item{type}{Types of output, which can be \strong{prob} (probability) and \strong{class} (predicted label). Default is \strong{class}.}

\item{...}{Additional arguments for predict method.}
}
\value{
Two types of output can be selected:
\item{prob}{Estimated probability of being a minority instance. The probability is averaged by using an equal-weight majority vote by all weak learners.}
\item{class}{Predicted class of the instance. Instances of probability larger than 0.5 are predicted as 1, otherwise 0.}
}
\description{
Predicting instances in test set using CSC4.5 object.
}
\examples{
data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- CSC45(x, y, pruning = TRUE)
output <- predict (model, x, type = "prob") # return probability estimation
output <- predict (model, x, type = "class") # return predicted class
}
