% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyEnsemble.R
\name{predict.EasyEnsemble}
\alias{predict.EasyEnsemble}
\title{Predict Method for EasyEnsemble Object}
\usage{
\method{predict}{EasyEnsemble}(object, x, type = "class", ...)
}
\arguments{
\item{object}{An object of EasyEnsemble class.}

\item{x}{A data frame of the predictors from testing data.}

\item{type}{Types of output, which can be \strong{probability} and \strong{class} (predicted label). Default is \strong{probability}.}

\item{...}{Additional arguments for predict method.}
}
\value{
Two types of output can be selected:
\item{probability}{Estimated probability of being a minority instance. The probability is averaged by using an equal-weight majority vote by all weak learners.}
\item{class}{Predicted class of the instance. Instances of probability larger than 0.5 are predicted as 1, otherwise 0.}
}
\description{
Predicting instances in test set using EasyEnsemble object.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
model <- EasyEnsemble(x, y, allowParallel=TRUE)
output <- predict (model, x, type = "probability") # return probability estimation
output <- predict (model, x, type = "class") # return predicted class
}
