% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomSampling.R
\name{RandomSampling}
\alias{RandomSampling}
\title{Implementation of Random Sampling Algorithm}
\usage{
RandomSampling(form, data, percOver = 0, percUnder = 6.8)
}
\arguments{
\item{form}{A model formula.}

\item{data}{A data frame of training data.}

\item{percOver}{Oversampling percentage.}

\item{percUnder}{Undersampling percentage.}
}
\value{
(newData) A data frame of the random oversampled/undersampled data.
}
\description{
This function implements random undersampling and oversampling algorithm.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
newData<- RandomSampling(Churn ~., trainset)
}
