% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTE.R
\name{SMOTE}
\alias{SMOTE}
\title{Synthetic Minority Oversampling Technique (SMOTE)}
\usage{
SMOTE(form, data, percOver = 1400, k = 5)
}
\arguments{
\item{form}{A model formula.}

\item{data}{A data frame of training data.}

\item{percOver}{Percent of new instance generated for each minority instance. percOver/100 is the number of new instances per minority point.}

\item{k}{Number of nearest neighbours.}
}
\value{
\item{newData}{A data frame of the oversampled data using SMOTE.}
}
\description{
This function implements SMOTE sampling (Synthetic Minority Oversampling Technique).
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
newData <- SMOTE(Churn ~., trainset)
}
\references{
Chawla, N., Bowyer, K., Hall, L. and Kegelmeyer, W. \emph{SMOTE: Synthetic minority oversampling technique.} Journal of Artificial Intelligence Research, 2002, 16(3), pp. 321-357.
}
