% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoteENN.R
\name{SmoteENN}
\alias{SmoteENN}
\title{SmoteENN - Implementation of SmoteENN Algorithm}
\usage{
SmoteENN(x, y, percOver = 1400, k1 = 5, k2 = 3, allowParallel = TRUE)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{percOver}{Percent of new instance generated for each minority instance.}

\item{k1}{Number of the nearest neighbors.}

\item{k2}{Number of neighbours for ENN.}

\item{allowParallel}{A logical number to control the parallel computing. If allowParallel = TRUE, the function is run using parallel techniques.}
}
\value{
\item{newData}{A data frame after the application of SmoteENN.}
}
\description{
This function implements SmoteENN algorithm, which combined SMOTE and data cleaning techniques ENN(Edited Nearest Neighbor).
}
