% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SmoteExs}
\alias{SmoteExs}
\title{SmoteExs}
\usage{
SmoteExs(data, percOver, k)
}
\arguments{
\item{data}{Dataset of the minority instances.}

\item{percOver}{Percentage of oversampling.}

\item{k}{number of nearest neighours.}
}
\value{
\item{newExs}{Dataframe with new instances.}
}
\description{
Obtain Smote instances for minority instances.
}
