% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoteTL.R
\name{TomekLink}
\alias{TomekLink}
\title{TomekLink}
\usage{
TomekLink(tgt, data)
}
\arguments{
\item{tgt}{Target columns index.}

\item{data}{Dataset.}
}
\value{
\item{indexTomek}{Logical vector indicating whether a instance is in TomekLinks.}
}
\description{
Function that finds the TomekLink
}
\keyword{internal}
