% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceBagging.R
\docType{data}
\name{treeBag}
\alias{treeBag}
\title{Basic tree for Bagging}
\format{
An object of class \code{list} of length 3.
}
\usage{
treeBag
}
\description{
Bagging Base learner
}
\keyword{internal}
