% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceBoost.R
\docType{data}
\name{treeBoost}
\alias{treeBoost}
\title{Basic tree for Boosting}
\format{
An object of class \code{list} of length 3.
}
\usage{
treeBoost
}
\description{
Boosting Base learner
}
\keyword{datasets}
