% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADASYN.R
\name{ADASYN}
\alias{ADASYN}
\title{Adaptive Synthetic Sampling}
\usage{
ADASYN(form, data, beta = 0.65, k = 5)
}
\arguments{
\item{form}{A model formula.}

\item{data}{A data frame of training data.}

\item{beta}{Balance level (0, 1], when beta=1, the dataset is fully balanced.}

\item{k}{Number of nearest neighbors}
}
\value{
\item{newData}{A data frame generated by ADASYN algorithm.}
}
\description{
This function implements ADASYN sampling.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
newData<- ADASYN(Churn ~., trainset, beta=0.8, k=5)
}
\references{
He, H., Y. Bai, et al. (2008). \emph{"ADASYN: Adaptive synthetic sampling approach for imbalanced learning."} IEEE Neural Networks(IJCNN)
}
