% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Korean}
\alias{Korean}
\title{Korean dataset}
\format{
A data frame with 26224 rows and 11 variables.
}
\usage{
data(Korean)
}
\description{
This dataset has 10 predictors with customers information and a binary column indicating whether client has churn or not.
}
\references{
Zhu, B., Gao, Z., Zhao, J., & Vanden Broucke, S. K. (2019). IRIC: An R library for binary imbalanced classification. SoftwareX, 10, 100341. https://doi.org/10.1016/j.softx.2019.100341
}
\author{
Unknown
}
\keyword{data}
\keyword{internal}
