% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomSampling.R
\name{RandomSampling}
\alias{RandomSampling}
\title{RandomSampling: Implementation of Random Sampling Algorithm}
\usage{
RandomSampling(x, y, percOver = 0, percUnder = 6.8)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{percOver}{Oversampling percentage.}

\item{percUnder}{Undersampling percentage.}
}
\value{
(newData) A data frame of the random oversampled/undersampled data.
}
\description{
This function implements random undersampling and oversampling algorithm.
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
newData<- RandomSampling(x, y)
}
