% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTE.R
\name{SMOTE}
\alias{SMOTE}
\title{Synthetic Minority Oversampling Technique (SMOTE)}
\usage{
SMOTE(x, y, percOver = 1400, k = 5)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{percOver}{Percent of new instance generated for each minority instance.}

\item{k}{Number of nearest neighbours.}
}
\value{
\item{newData}{A data frame of the oversampled data using SMOTE.}
}
\description{
This function implements SMOTE sampling (Synthetic Minority Oversampling Technique).
}
\examples{
data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
newData<- SMOTE(x, y)
}
\references{
Chawla, N., Bowyer, K., Hall, L. and Kegelmeyer, W. SMOTE: Synthetic minority oversampling technique. Journal of Artificial Intelligence Research, 2002, 16(3), pp. 321-357.
}
