% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoteENN.R
\name{SmoteENN}
\alias{SmoteENN}
\title{Implementation of SmoteENN Algorithm}
\usage{
SmoteENN(form, data, percOver = 1400, k1 = 5, k2 = 3, allowParallel = TRUE)
}
\arguments{
\item{form}{A model formula.}

\item{data}{A data frame of training data.}

\item{percOver}{Percent of new instance generated for each minority instance.}

\item{k1}{Number of the nearest neighbors.}

\item{k2}{Number of neighbours for ENN.}

\item{allowParallel}{A logical number to control the parallel computing. If allowParallel = TRUE, the function is run using parallel techniques.}
}
\value{
\item{newData}{A data frame after the application of SmoteENN.}
}
\description{
This function implements SmoteENN algorithm, which combined SMOTE and data cleaning techniques ENN(Edited Nearest Neighbor).
}
\examples{
library(caret)

data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
newData<- SmoteENN(Churn ~., trainset, percOver =1400, allowParallel = TRUE)
}
\references{
G. E. Batista, R. C. Prati, M. C. Monard. \emph{A study of the behavior of several methods for balancing machine learning training data.} ACM SIGKDD Explorations Newsletter , 6 (1) pp. 20 - 29.
}
