% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoteTL.R
\name{SmoteTL}
\alias{SmoteTL}
\title{SmoteTL - Smote sampling + TomekLinks}
\usage{
SmoteTL(x, y, percOver = 1400, k = 5)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{k}{Number of nearest neighbors used in Smote.}

\item{perc_over}{Number of new instance generated for each minority instance.}
}
\value{
\item{newData}{A data frame after the application of SmoteTL.}
}
\description{
This function implements SmoteTL, which performs over-sampling with SMOTE and clean data with Tomek Links.
}
\examples{
data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
newData<- SmoteTL(x, y, percOver = 1400)
}
\references{
G. E. Batista, R. C. Prati, M. C. Monard. A study of the behavior of several methods for balancing machine learning training data. ACM SIGKDD Explorations Newsletter , 6 (1) pp. 20 - 29.
}
