% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIDER.R
\name{SPIDER}
\alias{SPIDER}
\title{SPIDER}
\usage{
SPIDER(x, y, method = "weak", allowParallel = TRUE)
}
\arguments{
\item{x}{A data frame of the predictors from training data.}

\item{y}{A vector of response variable from training data.}

\item{method}{Type of modification of the minority class in the second phase, including "weak", "relabel", "strong".}

\item{allowParallel}{A logical number to control the parallel computing. If allowParallel = TRUE, the function is run using parallel techniques.}
}
\value{
\item{newData}{A data frame after the application of SPIDER.}
}
\description{
This function implements Perform SPIDER (Selective Preprocessing of Imbalanced Data with ENN Rule) on imbalanced dataset, which filters difficult instances from the majority class after local over sampling of the minority class.
}
\examples{
data(Korean)
sub <- createDataPartition(Korean$Churn,p=0.75,list=FALSE)
trainset <- Korean[sub,]
testset <- Korean[-sub,]
x <- trainset[, -11]
y <- trainset[, 11]
newData<- SPIDER(x, y, method = "weak", allowParallel= TRUE)
}
\references{
Stefanowski, J. and S. Wilk (2008). "Selective pre-processing of imbalanZced data for improving classification performance." Data Warehousing and Knowledge Discovery, Springer: 283-292.
}
